refactor: Add /legacy export path
