+++ packages/json-rpc-engine/src/index.test.ts
@@ -4,12 +4,7 @@ describe('@metamask/json-rpc-engine', () => {
   it('has expected JavaScript exports', () => {
     expect(Object.keys(allExports)).toMatchInlineSnapshot(`
       Array [
-        "createAsyncMiddleware",
-        "createScaffoldMiddleware",
         "getUniqueId",
-        "createIdRemapMiddleware",
-        "JsonRpcEngine",
-        "mergeMiddleware",
       ]
     `);
   });
