+++ packages/json-rpc-engine/src/index.ts
@@ -1,19 +1,1 @@
-// Legacy
-export type {
-  AsyncJsonRpcEngineNextCallback,
-  AsyncJsonrpcMiddleware,
-} from './legacy/createAsyncMiddleware';
-export { createAsyncMiddleware } from './legacy/createAsyncMiddleware';
-export { createScaffoldMiddleware } from './legacy/createScaffoldMiddleware';
 export { getUniqueId } from './getUniqueId';
-export { createIdRemapMiddleware } from './legacy/idRemapMiddleware';
-export type {
-  JsonRpcEngineCallbackError,
-  JsonRpcEngineReturnHandler,
-  JsonRpcEngineNextCallback,
-  JsonRpcEngineEndCallback,
-  JsonRpcMiddleware,
-  JsonRpcNotificationHandler,
-} from './legacy/JsonRpcEngine';
-export { JsonRpcEngine } from './legacy/JsonRpcEngine';
-export { mergeMiddleware } from './legacy/mergeMiddleware';
