+++ packages/json-rpc-engine/src/legacy/index.ts
@@ -1,0 +1,17 @@
+export type {
+  AsyncJsonRpcEngineNextCallback,
+  AsyncJsonrpcMiddleware,
+} from './createAsyncMiddleware';
+export { createAsyncMiddleware } from './createAsyncMiddleware';
+export { createScaffoldMiddleware } from './createScaffoldMiddleware';
+export { createIdRemapMiddleware } from './idRemapMiddleware';
+export type {
+  JsonRpcEngineCallbackError,
+  JsonRpcEngineReturnHandler,
+  JsonRpcEngineNextCallback,
+  JsonRpcEngineEndCallback,
+  JsonRpcMiddleware,
+  JsonRpcNotificationHandler,
+} from './JsonRpcEngine';
+export { JsonRpcEngine } from './JsonRpcEngine';
+export { mergeMiddleware } from './mergeMiddleware';
